move type guards into folder
