+++ libs/extensions-api/create-extension-element-or-fallback.function.ts
@@ -1,7 +1,10 @@
 import { createExtensionElement } from './create-extension-element.function';
-import { isManifestElementableType } from './is-manifest-elementable-type.function';
+import { isManifestElementableType } from './type-guards';
 
-export async function createExtensionElementOrFallback(manifest: any, fallbackElementName: string): Promise<HTMLElement | undefined> {
+export async function createExtensionElementOrFallback(
+	manifest: any,
+	fallbackElementName: string
+): Promise<HTMLElement | undefined> {
 	if (isManifestElementableType(manifest)) {
 		return createExtensionElement(manifest);
 	}
