+++ libs/extensions-api/type-guards/index.ts
@@ -1,0 +1,6 @@
+export * from './is-manifest-class-instance-type.function';
+export * from './is-manifest-classable-type.function';
+export * from './is-manifest-element-name-type.function';
+export * from './is-manifest-elementable-type.function';
+export * from './is-manifest-js-type.function';
+export * from './is-manifest-loader-type.function';
