+++ libs/extensions-api/type-guards/is-manifest-class-instance-type.function.ts
