+++ libs/extensions-api/type-guards/is-manifest-element-name-type.function.ts
