+++ libs/extensions-api/type-guards/is-manifest-elementable-type.function.ts
@@ -3,7 +3,9 @@ import { isManifestJSType } from './is-manifest-js-type.function';
 import { isManifestLoaderType } from './is-manifest-loader-type.function';
 import type { ManifestElement, ManifestBase } from '@umbraco-cms/backoffice/extensions-registry';
 
-export function isManifestElementableType<ElementType extends HTMLElement = HTMLElement>(manifest: ManifestBase): manifest is ManifestElement {
+export function isManifestElementableType<ElementType extends HTMLElement = HTMLElement>(
+	manifest: ManifestBase
+): manifest is ManifestElement {
 	return (
 		isManifestElementNameType(manifest) ||
 		isManifestLoaderType<ElementType>(manifest) ||
