+++ libs/extensions-api/type-guards/is-manifest-js-type.function.ts
