+++ libs/extensions-api/type-guards/is-manifest-loader-type.function.ts
